refactor(json-schema-2020-12): make styles more reusable
Refs #8513
