+++ src/core/plugins/json-schema-2020-12/components/keywords/AdditionalProperties.jsx
@@ -27,12 +27,16 @@ const AdditionalProperties = ({ schema }) => {
       {additionalProperties === true ? (
         <>
           {name}
-          <span className="json-schema-2020-12__type">allowed</span>
+          <span className="json-schema-2020-12__attribute json-schema-2020-12__attribute--primary">
+            allowed
+          </span>
         </>
       ) : additionalProperties === false ? (
         <>
           {name}
-          <span className="json-schema-2020-12__type">forbidden</span>
+          <span className="json-schema-2020-12__attribute json-schema-2020-12__attribute--primary">
+            forbidden
+          </span>
         </>
       ) : (
         <JSONSchema name={name} schema={additionalProperties} />
