+++ src/core/plugins/json-schema-2020-12/components/keywords/Deprecated.jsx
@@ -3,12 +3,16 @@
  */
 import React from "react"
 
-import { schema } from "../../../prop-types"
+import { schema } from "../../prop-types"
 
 const Deprecated = ({ schema }) => {
   if (schema?.deprecated !== true) return null
 
-  return <span className="json-schema-2020-12__deprecated">deprecated</span>
+  return (
+    <span className="json-schema-2020-12__attribute json-schema-2020-12__attribute--warning">
+      deprecated
+    </span>
+  )
 }
 
 Deprecated.propTypes = {
