--- src/core/plugins/json-schema-2020-12/components/keywords/Deprecated/_deprecated.scss
@@ -1,10 +1,0 @@
-.json-schema-2020-12 {
-  &__deprecated {
-    @include text_code();
-    color: red;
-    font-size: 12px;
-    text-transform: lowercase;
-    font-weight: normal;
-    padding-left: 10px;
-  }
-}
