+++ src/core/plugins/json-schema-2020-12/components/keywords/ReadOnly.jsx
@@ -3,12 +3,16 @@
  */
 import React from "react"
 
-import { schema } from "../../../prop-types"
+import { schema } from "../../prop-types"
 
 const ReadOnly = ({ schema }) => {
   if (schema?.readOnly !== true) return null
 
-  return <span className="json-schema-2020-12__readOnly">read-only</span>
+  return (
+    <span className="json-schema-2020-12__attribute json-schema-2020-12__attribute--muted">
+      read-only
+    </span>
+  )
 }
 
 ReadOnly.propTypes = {
