--- src/core/plugins/json-schema-2020-12/components/keywords/ReadOnly/_read-only.scss
@@ -1,6 +1,0 @@
-.json-schema-2020-12 {
-  &__readOnly {
-    @extend .json-schema-2020-12__deprecated;
-    color: gray;
-  }
-}
