+++ src/core/plugins/json-schema-2020-12/components/keywords/Type.jsx
@@ -4,8 +4,8 @@
 import React from "react"
 import PropTypes from "prop-types"
 
-import { schema } from "../../../prop-types"
-import { useFn } from "../../../hooks"
+import { schema } from "../../prop-types"
+import { useFn } from "../../hooks"
 
 const Type = ({ schema, isCircular }) => {
   const fn = useFn()
@@ -13,7 +13,9 @@ const Type = ({ schema, isCircular }) => {
   const circularSuffix = isCircular ? " [circular]" : ""
 
   return (
-    <span className="json-schema-2020-12__type">{`${type}${circularSuffix}`}</span>
+    <strong className="json-schema-2020-12__attribute json-schema-2020-12__attribute--primary">
+      {`${type}${circularSuffix}`}
+    </strong>
   )
 }
