--- src/core/plugins/json-schema-2020-12/components/keywords/Type/_type.scss
@@ -1,18 +1,0 @@
-.json-schema-2020-12 {
-  &__type {
-    @include text_code();
-    color: $prop-type-font-color;
-    font-size: 12px;
-    text-transform: lowercase;
-    font-weight: bold;
-    padding-left: 10px;
-  }
-
-  &-property {
-    .json-schema-2020-12__type {
-      vertical-align: middle;
-    }
-  }
-}
-
-
