+++ src/core/plugins/json-schema-2020-12/components/keywords/WriteOnly.jsx
@@ -3,12 +3,16 @@
  */
 import React from "react"
 
-import { schema } from "../../../prop-types"
+import { schema } from "../../prop-types"
 
 const WriteOnly = ({ schema }) => {
   if (schema?.writeOnly !== true) return null
 
-  return <span className="json-schema-2020-12__writeOnly">write-only</span>
+  return (
+    <span className="json-schema-2020-12__attribute json-schema-2020-12__attribute--muted">
+      write-only
+    </span>
+  )
 }
 
 WriteOnly.propTypes = {
