--- src/core/plugins/json-schema-2020-12/components/keywords/WriteOnly/_write-only.scss
@@ -1,5 +1,0 @@
-.json-schema-2020-12 {
-  &__writeOnly {
-    @extend .json-schema-2020-12__readOnly;
-  }
-}
