+++ src/core/plugins/json-schema-2020-12/index.js
@@ -28,7 +28,7 @@ import KeywordAdditionalProperties from "./components/keywords/AdditionalPropert
 import KeywordPropertyNames from "./components/keywords/PropertyNames"
 import KeywordUnevaluatedItems from "./components/keywords/UnevaluatedItems"
 import KeywordUnevaluatedProperties from "./components/keywords/UnevaluatedProperties"
-import KeywordType from "./components/keywords/Type/Type"
+import KeywordType from "./components/keywords/Type"
 import KeywordEnum from "./components/keywords/Enum/Enum"
 import KeywordConst from "./components/keywords/Const"
 import KeywordConstraint from "./components/keywords/Constraint/Constraint"
@@ -37,9 +37,9 @@ import KeywordContentSchema from "./components/keywords/ContentSchema"
 import KeywordTitle from "./components/keywords/Title/Title"
 import KeywordDescription from "./components/keywords/Description/Description"
 import KeywordDefault from "./components/keywords/Default"
-import KeywordDeprecated from "./components/keywords/Deprecated/Deprecated"
-import KeywordReadOnly from "./components/keywords/ReadOnly/ReadOnly"
-import KeywordWriteOnly from "./components/keywords/WriteOnly/WriteOnly"
+import KeywordDeprecated from "./components/keywords/Deprecated"
+import KeywordReadOnly from "./components/keywords/ReadOnly"
+import KeywordWriteOnly from "./components/keywords/WriteOnly"
 import Accordion from "./components/Accordion/Accordion"
 import ExpandDeepButton from "./components/ExpandDeepButton/ExpandDeepButton"
 import ChevronRightIcon from "./components/icons/ChevronRight"
